Clean up the plugin area in view of transition to dd4hep
