+++ Geometry/HGCalCommonData/plugins/DDHGCalNoTaperEndcap.cc
@@ -1,17 +1,52 @@
 #include <algorithm>
 #include <cmath>
+#include <string>
+#include <vector>
 
 #include "DataFormats/Math/interface/GeantUnits.h"
+#include "DetectorDescription/Core/interface/DDAlgorithm.h"
+#include "DetectorDescription/Core/interface/DDAlgorithmFactory.h"
 #include "DetectorDescription/Core/interface/DDCurrentNamespace.h"
 #include "DetectorDescription/Core/interface/DDLogicalPart.h"
 #include "DetectorDescription/Core/interface/DDMaterial.h"
 #include "DetectorDescription/Core/interface/DDSolid.h"
 #include "DetectorDescription/Core/interface/DDSplit.h"
+#include "DetectorDescription/Core/interface/DDTypes.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
-#include "Geometry/HGCalCommonData/plugins/DDHGCalNoTaperEndcap.h"
+#include "FWCore/PluginManager/interface/PluginFactory.h"
 
 //#define EDM_ML_DEBUG
 using namespace geant_units::operators;
+
+class DDHGCalNoTaperEndcap : public DDAlgorithm {
+public:
+  DDHGCalNoTaperEndcap(void);
+  ~DDHGCalNoTaperEndcap(void) override;
+
+  void initialize(const DDNumericArguments& nArgs,
+                  const DDVectorArguments& vArgs,
+                  const DDMapArguments& mArgs,
+                  const DDStringArguments& sArgs,
+                  const DDStringVectorArguments& vsArgs) override;
+
+  void execute(DDCompactView& cpv) override;
+
+private:
+  int createQuarter(DDCompactView& cpv, int xQuadrant, int yQuadrant, int startCopyNo);
+
+  double m_startAngle;        // Start angle
+  double m_tiltAngle;         // Tilt  angle
+  int m_invert;               // Inverted or forward
+  double m_rMin;              // Inner radius
+  double m_rMax;              // Outer radius
+  double m_zoffset;           // Offset in z
+  double m_xyoffset;          // Offset in x or y
+  int m_n;                    // Mumber of copies
+  int m_startCopyNo;          // Start copy Number
+  int m_incrCopyNo;           // Increment copy Number
+  std::string m_childName;    // Children name
+  std::string m_idNameSpace;  // Namespace of this and ALL sub-parts
+};
 
 DDHGCalNoTaperEndcap::DDHGCalNoTaperEndcap() {
   edm::LogVerbatim("HGCalGeom") << "DDHGCalNoTaperEndcap test: Creating an instance";
@@ -132,3 +167,5 @@ int DDHGCalNoTaperEndcap::createQuarter(DDCompactView& cpv, int xQuadrant, int y
 #endif
   return copyNo;
 }
+
+DEFINE_EDM_PLUGIN(DDAlgorithmFactory, DDHGCalNoTaperEndcap, "hgcal:DDHGCalNoTaperEndcap");
