+++ Geometry/HGCalCommonData/plugins/DDHGCalWafer8.cc
@@ -1,13 +1,44 @@
-#include "Geometry/HGCalCommonData/plugins/DDHGCalWafer8.h"
+#include "DetectorDescription/Core/interface/DDAlgorithm.h"
+#include "DetectorDescription/Core/interface/DDAlgorithmFactory.h"
 #include "DetectorDescription/Core/interface/DDCurrentNamespace.h"
 #include "DetectorDescription/Core/interface/DDLogicalPart.h"
 #include "DetectorDescription/Core/interface/DDMaterial.h"
 #include "DetectorDescription/Core/interface/DDSolid.h"
 #include "DetectorDescription/Core/interface/DDSplit.h"
+#include "DetectorDescription/Core/interface/DDTypes.h"
 #include "DetectorDescription/Core/interface/DDutils.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/PluginManager/interface/PluginFactory.h"
+
+#include <string>
+#include <vector>
 
 //#define EDM_ML_DEBUG
+
+class DDHGCalWafer8 : public DDAlgorithm {
+public:
+  // Constructor and Destructor
+  DDHGCalWafer8();
+  ~DDHGCalWafer8() override;
+
+  void initialize(const DDNumericArguments& nArgs,
+                  const DDVectorArguments& vArgs,
+                  const DDMapArguments& mArgs,
+                  const DDStringArguments& sArgs,
+                  const DDStringVectorArguments& vsArgs) override;
+  void execute(DDCompactView& cpv) override;
+
+private:
+  double waferSize_;                    // Wafer size
+  double waferT_;                       // Wafer thickness
+  double waferSepar_;                   // Sensor separation
+  double mouseBite_;                    // MouseBite radius
+  int nCells_;                          // Half number of cells along u-v axis
+  int cellType_;                        // Cell Type (0,1,2: Fine, Course 2/3)
+  std::string material_;                // Material name for module with gap
+  std::vector<std::string> cellNames_;  // Name of the cells
+  std::string nameSpace_;               // Namespace to be used
+};
 
 DDHGCalWafer8::DDHGCalWafer8() {
 #ifdef EDM_ML_DEBUG
@@ -115,3 +146,5 @@ void DDHGCalWafer8::execute(DDCompactView &cpv) {
     }
   }
 }
+
+DEFINE_EDM_PLUGIN(DDAlgorithmFactory, DDHGCalWafer8, "hgcal:DDHGCalWafer8");
