+++ Geometry/HGCalCommonData/python/testGeometryExtended_cff.py
@@ -1,6 +1,6 @@
 import FWCore.ParameterSet.Config as cms
 
-from Geometry.CMSCommonData.cmsExtendedGeometry2026D28XML_cfi import *
+from Geometry.CMSCommonData.cmsExtendedGeometry2026D41XML_cfi import *
 from Geometry.TrackerNumberingBuilder.trackerNumberingGeometry_cfi import *
 from Geometry.HcalCommonData.hcalDDConstants_cff import *
 from Geometry.HGCalCommonData.hgcalParametersInitialization_cfi import *
