+++ Geometry/HGCalCommonData/test/dumpExtendedHGCalGeometry_cfg.py
